move date formatting function to evutil_time.c.
